// src/interface/components/common.interface.ts
export enum AlignmentEnum {
  LEFT = "left",
  RIGHT = "right",
}

export enum ColorEnum {
  RED = "red",
  GREEN = "green",
  BLUE = "blue",
  YELLOW = "yellow",
  ORANGE = "orange",
  PURPLE = "purple",
  PINK = "pink",
  BROWN = "brown",
  BLACK = "black",
  WHITE = "white",
  GREY = "grey",
  TRANSPARENT = "transparent",
  PRIMARY = "primary",
  SECONDARY = "secondary",
  ERROR = "error",
  WARNING = "warning",
  INFO = "info",
}